refactor(datepicker): extract shared PickDateAdapter and GIDDH_DATEPICKER_FORMAT
- Created giddh-date-adapter/giddh-date-adapter.ts as single source of truth
- Removed ~550 lines of duplication from both datepicker modules
- Added compact no-space date parsing (12jan25, 12j23, 12jul)
- Added no-year parsing defaulting to current year (12jan, 12j, jan12)
- Added prefix-based month matching (j→Jan, ju→Jun, d→Dec)
