+++ apps/web-giddh/src/app/vouchers/create/create.component.html
@@ -1734,7 +1734,7 @@
                                                                 <div class="mt-2 font-size-12">
                                                                     <span>{{ localeData?.entry_date }}</span>
                                                                     <div class="mt-1" *ngIf="activeEntryIndex !== entryIndex">
-                                                                        {{ entry.get('date')?.value }}
+                                                                        {{ entry.get('date')?.value | giddhDate }}
                                                                     </div>
                                                                     <div
                                                                         class="mt-1 d-flex align-items-baseline column-gap-05"
